pyproject.toml
[tool.poetry]
name = "py-gmail"
version = "0.1.0"
description = ""
authors = ["Harshad Parulekar <harshadhparulekar@gmail.com>"]


[tool.poetry.dependencies]
python = "^3.10"
python-decouple = "^3.8"
html2text = "^2020.1.16"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
